Setup property profiles for different environments
Externalize application-local, -prod, and -test properties files, which
contain connection information including passwords and keys. These files
are not added to source control.  The application.properties file no
longer contains this information, but is used to contain general
properties and to set the environment profile (test, prod, local).
That setting tells the Spring to look for the specific properties file.